requirements.txt
requests
langchain
langchain_openai
openai
faiss-cpu
fastapi
uvicorn
python-dotenv
